All parser tests pass with FLYING COLORS
I organized the parser tests into subsections because there are a lot of
them now. I also added the 'continue' keyword to the grammar because you
sort of need it if you have 'break'. The entire parser file has 100%
coverage, and the logic has been adjusted slightly to make that happen.
Now we can move on to type checking!