Added tons of new logic, made some decisions around type checking
So I've decided how the passes are going to work.
Pass 1 is the declaration processing pass, where we take all top-level
declarations in every module and add them to a corresponding table in the
module. Every module-scoped name is made available after this pass.
Pass 2 is the type resolution pass, where every declaration is visited to
determine its type, and every expression and statement is visited along
with an expected type to determine the semantic validity of the program.
Pass 2 was originally going to be two passes: a type processing pass, and
then a type checking pass. I now realize that if we resolve names as we
see them instead of one at a time, these steps can be done in parallel.
Pass 1 has been completed for a few days, pass 2 is now complete for all
declarations, types, and statements. I still have yet to figure out some
of the more complicated expressions. All typecheck type classes now have
an isAssignableFrom() method that determines if another type is assignable
to itself, but I'm finding that I'm not using it very much because the
logic is being reimplemented in the resolveType() methods in the
expression node classes. I need to figure out how to combine this logic
with the error handling, but at the same time handle cases where the
expected type is only partial, for example with the iterable expression in
for loops. Once I figure that out I can implement the rest of the
resolveType() methods and type checking will be ready for unit testing.