Added operators, finished type resolution logic for all nodes
Operators ended up being just a lot of code, and a lot of complexity.
Operators have a type (prefix, postfix, infix), a symbol, and (if infix) a
precedence and associativity. The latter two values determine what the
tree structure will look like. When we are resoling a binary expression,
we also need to resolve the precedence, so an initial version of that
logic was added.

Additionally, the logic for inferring lambda types has also been added.
Kinda sucks that function application is the only place we can use lambdas
for now, but that's just the way it has to be.

With those two above points, the hairiest remaining aspects of this
process are complete and all type resolution logic is done. I found
several places along the way that are certain to be broken, but the logic
is at least 95% done at this point.

Next steps are to run through it manually one more time, then start
writing tests. We will try to work from the ground up.