Nearly complete with translator logic, added interpreter logic
So, it's difficult to tell exactly what the translator should do without
knowing what the interpreter will do. So I got as far as I could with the
translator and then implemented the interpreter. Turns out for everything
except return, throw, and break/continue, it was all super easy. But now I
have a better picture of how this should work. Once I finalize this logic,
I can test it :D and then we can get on implementing the operators and
some base standard library functions. Then we should be good to get going
on building the compiler!