Translator process is now fully tested
All that is left is the interpreter. However, it will be difficult to
properly test it without having at least a small standard library.
So what I'm working on now is sketching out exactly what will be included
in it. Generics and overloads (plus classes and extensions) are coming up
a lot, so I may need to start implementing those before starting on the
standard library.