Refactored parser stack to use immutable lazy lists
We were running into several problems related to the complex tangle of state,
and at the core of the issue was the fact that our LookaheadIterator is dependent
on severely mutable state. I thus created an immutable equivalent called
LazyList, which provides the exact behavior of the LookaheadIterator, but in
an immutable way, and is ironically far simpler. The class is backed by normal
iteration logic, so not much had to be changed to support it.

Once that was done, the parser control logic was refactored to use it, which
ended up in a much cleaner and easier-to-understand process that I was able
to split into several methods. I don't imagine we will have many other problems
from here on out, and if we do, they should be very simple to solve.