Finished all test for the new parser structure, removed all the old
So that was probably the biggest step.
Since we tweaked the import syntax slightly, I also changed all now-incorrect
instances.
Now begins some more tedious work. We need to fix everywhere where we changed
the names of components, and, more importantly, changed the AST structure
to support the new parser format. There are going to be a lot more reduce()
functions, and they should try to keep the reduced structure as close to
how it was as possible. There shouldn't be much issue here, just more logic.

Once this is done and all tests are passing, the real fun can begin.
We can finally start adding our new syntax features that are required
for the v0 compiler.