Converted tests to TypeScript
This was a loooong process, but we came out with the following benefits:
- An actual scripting setup for building and testing
- Tests building in TS
- The beginnings of some test utilities
- We can debug tests in VS Code!
Now everything is in TypeScript (except the scripts) so yay!
Some things that still need to be done:
- Fix 22 remaining failing tests
- Clean up the tests so that:
  - Everything uses chai.assert instead of chai.expect
  - There is waaaay less duplicated code
  - We have a common set of test utilities
  - Coverage is working
  - Coverage is back to old numbers
For now, I want to focus on actual feature work, so tests will take
the back burner for awhile.