All tests pass, coverage works
The last vestiges of the old world are now gone, including:
- .babelrc
- .eslintrc
- webstormfixes.js

The tests pass. This required some small tweaks in a few places,
as well as the slightly larger change of adding a new TInferred
type to replace the usage of 'null' for inferring param/return types
in lambda functions.

Coverage works now, just like it did before, but we did need to add
an istanbul config file to keep some files ignored.