Migrated to new parser mechanism
When adding new parser logic and writing parser tests, I noticed that it
was difficult and over-complicated, so I decided to reimagine the logic
to simplify it and make it easier to add new stuff later.

Previously, the parser had 4 parts:
- the grammar definition (parser/parser-impl.ts)
- the CST nodes (syntax/*/cst.ts)
- the reducer functions (syntax/*/reducer.ts)
- the AST nodes (syntax/*/ast.ts)
The grammar definition and reducer functions are necessary logic, but
they can be combined. The CST nodes were purely intermediate and could be removed.
All we care about in the end is the AST nodes, so I found a way to merge all four
into one.

The grammar definition is now specified by a set of decorators on the
AST nodes. The CST nodes are the target methods of the decorators,
which contain the reducing logic and set the AST node properties.
That way the logic all exists in the AST node. You can now parse an
AST node by passing the class into the parser. This makes adding parser
logic and testing it much simpler.

Additionally, I added the following changes:
- Token types are now done via an enum.
- Reserved words now have a specific token type instead of
  one type for each word.
- Negative numbers are no longer individual tokens because
  there is ambiguity with binary expressions with no spaces.